// src/shared/types/button.ts
import { ButtonHTMLAttributes } from "react";

export type CloseButtonProps = {
  onClick: () => void;
};

export interface ButtonProps {
  children: React.ReactNode;
  onClick: () => void;
  variant?: "primary" | "secondary" | "google";
  type?: "submit" | "button";
  disabled?: any;
}

export interface IconButtonProps
  extends ButtonHTMLAttributes<HTMLButtonElement> {
  children: React.ReactNode;
  "aria-label": string;
}

export interface OptionButtonProps {
  icon: React.ElementType;
  text: string;
  onClick: () => void;
}

export interface CategoryButtonProps {
  label: string;
  onClick?: () => void;
}

export interface UploadProps {
  onUploadClick: () => void;
}

export type DocumentType = "image" | "document" | "transaction";

export interface Document {
  type: DocumentType;
  name: string;
  description?: string;
  amount?: number;
  category?: string;
  timestamp?: string;
  isIncome?: boolean;
  color?: string;
}

export interface SentDocumentsSectionProps {
  documents: Document[];
  showOptions: boolean;
  onFileSelect: (file: File, type: "image" | "document") => void;
  onManualClick?: () => void;
}

export interface UploadOptionsProps {
  onFileSelect: (file: File, type: "image" | "document") => void;
  onManualClick?: () => void;
}

export interface NavigationBackProps {
  prev?: () => void;
  color?: string;
}

export interface MoreButtonProps {
  onClick?: () => void;
  isLoading?: boolean;
  color?: string;
  label: string;
}

export interface OptionChatButtonProps {
  label: string;
}

export interface LandingButtonProps {
  children: React.ReactNode;
  onClick?: () => void;
  variant?: "primary" | "secondary";
  className?: string;
  type?: "button" | "submit" | "reset";
}